refactor: remove unused exports cy-314
