+++ apps/frontend/src/libs/components/components.ts
@@ -13,6 +13,5 @@ export { Logo } from "./logo/logo.js";
 export { Modal } from "./modal/modal.js";
 export { ProtectedRoute } from "./protected-route/protected-route.js";
 export { RouterProvider } from "./router-provider/router-provider.js";
-export { TabsWithConfirmation } from "./tabs/tabs-with-confirmation.js";
 export { Tabs } from "./tabs/tabs.js";
 export { Provider as StoreProvider } from "react-redux";
